requirements.txt
alembic==1.8.1
anyio==3.6.2
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.85.1
fastapi-pagination==0.10.0
FastAPI-SQLAlchemy==0.2.1
flake8==5.0.4
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.42
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
zipp==3.10.0
